Cleanup graph-related code.
