--- src/main/kotlin/com/autonomousapps/graph/DepthFirstSearch.kt
@@ -1,40 +1,0 @@
-package com.autonomousapps.graph
-
-/**
- * This class is used to generate subgraphs from a larger graph, [graph], which are reachable from
- * the given source node.
- */
-internal class DepthFirstSearch(private val graph: DependencyGraph, private val source: Node) {
-
-  constructor(graph: DependencyGraph, source: String) : this(graph, BareNode(source))
-
-  private val marked = linkedMapOf<String, Boolean>()
-
-  init {
-    dfs(source)
-  }
-
-  private fun dfs(source: Node) {
-    marked[source.identifier] = true
-    graph.adj(source).forEach { (_, to) ->
-      if (!isMarked(to)) {
-        dfs(to)
-      }
-    }
-  }
-
-  val subgraph: DependencyGraph by lazy {
-    val sub = DependencyGraph()
-    graph.edges().forEach {
-      // if both nodes on the edge are marked, add this edge to the subgraph
-      if (isMarked(it.from) && isMarked(it.to)) {
-        sub.addEdge(it)
-      }
-    }
-    sub.addNode(source)
-    sub
-  }
-
-  private fun isMarked(node: Node): Boolean = isMarked(node.identifier)
-  private fun isMarked(identifier: String): Boolean = marked.getOrDefault(identifier, false)
-}
