+++ src/main/kotlin/com/autonomousapps/tasks/GraphViewTask.kt
@@ -3,24 +3,19 @@ package com.autonomousapps.tasks
 import com.autonomousapps.TASK_GROUP_DEP_INTERNAL
 import com.autonomousapps.internal.GradleVersions
 import com.autonomousapps.internal.artifactsFor
-import com.autonomousapps.internal.isJavaPlatform
-import com.autonomousapps.internal.utils.*
-import com.autonomousapps.model.Coordinates
+import com.autonomousapps.internal.graph.GraphViewBuilder
+import com.autonomousapps.internal.graph.GraphWriter
+import com.autonomousapps.internal.utils.getAndDelete
+import com.autonomousapps.internal.utils.toJson
 import com.autonomousapps.model.DependencyGraphView
-import com.autonomousapps.model.ProjectCoordinates
 import com.autonomousapps.model.declaration.SourceSetKind
 import com.autonomousapps.model.declaration.Variant
-import com.google.common.graph.Graph
 import org.gradle.api.DefaultTask
 import org.gradle.api.artifacts.Configuration
-import org.gradle.api.artifacts.FileCollectionDependency
-import org.gradle.api.artifacts.result.ResolvedComponentResult
-import org.gradle.api.artifacts.result.ResolvedDependencyResult
 import org.gradle.api.file.FileCollection
 import org.gradle.api.file.RegularFileProperty
 import org.gradle.api.provider.Property
 import org.gradle.api.tasks.*
-import org.gradle.kotlin.dsl.support.appendReproducibleNewLine
 
 @CacheableTask
 abstract class GraphViewTask : DefaultTask() {
@@ -82,102 +77,5 @@ abstract class GraphViewTask : DefaultTask() {
 
     output.writeText(graphView.toJson())
     outputDot.writeText(GraphWriter.toDot(graph))
-  }
-}
-
-/** Walks the resolved dependency graph to create a dependency graph rooted on the current project. */
-@Suppress("UnstableApiUsage") // Guava Graph
-private class GraphViewBuilder(conf: Configuration) {
-
-  val graph: Graph<Coordinates>
-
-  private val graphBuilder = DependencyGraphView.newGraphBuilder()
-
-  private val visited = mutableSetOf<Coordinates>()
-
-  init {
-    val root = conf
-      .incoming
-      .resolutionResult
-      .root
-
-    val rootId = conf.rootCoordinates()
-
-    walkFileDeps(conf, rootId)
-    walk(root, rootId)
-
-    graph = graphBuilder.build()
-  }
-
-  private fun walkFileDeps(conf: Configuration, rootId: Coordinates) {
-    graphBuilder.addNode(rootId)
-
-    // the only way to get flat jar file dependencies
-    conf.allDependencies
-      .filterIsInstance<FileCollectionDependency>()
-      .mapNotNullToSet { it.toCoordinates() }
-      .forEach { id ->
-        graphBuilder.putEdge(rootId, id)
-      }
-  }
-
-  private fun walk(root: ResolvedComponentResult, rootId: Coordinates) {
-    root.dependencies
-      .filterIsInstance<ResolvedDependencyResult>()
-      // AGP adds all runtime dependencies as constraints to the compile classpath, and these show
-      // up in the resolution result. Filter them out.
-      .filterNot { it.isConstraint }
-      // For similar reasons as above
-      .filterNot { it.isJavaPlatform() }
-      // Sometimes there is a self-dependency?
-      .filterNot { it.selected == root }
-      .forEach { dependencyResult ->
-        // Might be from an included build, in which case the coordinates reflect the _requested_ dependency instead of
-        // the _resolved_ dependency.
-        val depId = dependencyResult.toCoordinates()
-
-        // add an edge
-        graphBuilder.putEdge(rootId, depId)
-
-        if (!visited.contains(depId)) {
-          visited.add(depId)
-          // recursively walk the graph in a depth-first pattern
-          walk(dependencyResult.selected, depId)
-        }
-      }
-  }
-}
-
-// TODO move
-@Suppress("UnstableApiUsage")
-internal object GraphWriter {
-
-  fun toDot(graph: Graph<Coordinates>) = buildString {
-    val projectNodes = graph.nodes()
-      .filterIsInstance<ProjectCoordinates>()
-      .map { it.gav() }
-
-    appendReproducibleNewLine("strict digraph DependencyGraph {")
-    appendReproducibleNewLine("  ratio=0.6;")
-    appendReproducibleNewLine("  node [shape=box];")
-
-    // styling for project nodes
-    if (projectNodes.isNotEmpty()) appendReproducibleNewLine()
-    projectNodes.forEach {
-      appendReproducibleNewLine("  \"$it\" [style=filled fillcolor=\"#008080\"];")
-    }
-    if (projectNodes.isNotEmpty()) appendReproducibleNewLine()
-
-    // the graph itself
-    graph.edges().forEach { edge ->
-      val source = edge.nodeU()
-      val target = edge.nodeV()
-      val style =
-        if (source is ProjectCoordinates && target is ProjectCoordinates) " [style=bold color=\"#FF6347\" weight=8]"
-        else ""
-      append("  \"${source.gav()}\" -> \"${target.gav()}\"$style;")
-      append("\n")
-    }
-    append("}")
   }
 }
